+++ projects/example/src/app/app.routes.ts
@@ -42,6 +42,8 @@ export const routes: Routes = [
   { path: 'pk-barcode', loadComponent: () => import('./pages/pk-barcode/pk-barcode-page').then(m => m.PkBarcodePage) },
   { path: 'pk-qrcode', loadComponent: () => import('./pages/pk-qrcode/pk-qrcode-page').then(m => m.PkQrcodePage) },
   { path: 'pk-code-reader', loadComponent: () => import('./pages/pk-code-reader/pk-code-reader-page').then(m => m.PkCodeReaderPage) },
+  { path: 'pk-radio-group', loadComponent: () => import('./pages/pk-radio-group/pk-radio-group-page').then(m => m.PkRadioGroupPage) },
+  { path: 'pk-timepicker', loadComponent: () => import('./pages/pk-timepicker/pk-timepicker-page').then(m => m.PkTimepickerPage) },
   // Examples
   { path: 'examples/login', loadComponent: () => import('./pages/examples/login/login-example').then(m => m.LoginExample) },
   { path: 'examples/chat', loadComponent: () => import('./pages/examples/chat/chat-example').then(m => m.ChatExample) },